+++ es/core/helper/set-global-state.js
@@ -1,5 +1,5 @@
 import util from '../../support/util';
-import { BROADCAST_TRIGGERED_BY_CC_API_SET_GLOBAL_STATE } from '../../support/constant';
+import { BROADCAST_TRIGGERED_BY_CC_API_SET_GLOBAL_STATE, MODULE_GLOBAL } from '../../support/constant';
 import pickOneRef from './pick-one-ref';
 /****
  * if you are sure this state is really belong to global state, call cc.setGlobalState,
@@ -13,7 +13,11 @@ export default function (state, throwError) {
 
   try {
     var ref = pickOneRef();
-    ref.setGlobalState(state, BROADCAST_TRIGGERED_BY_CC_API_SET_GLOBAL_STATE);
+    ref.$$changeState(state, {
+      module: MODULE_GLOBAL,
+      stateFor: STATE_FOR_ALL_CC_INSTANCES_OF_ONE_MODULE,
+      broadcastTriggeredBy: BROADCAST_TRIGGERED_BY_CC_API_SET_GLOBAL_STATE
+    });
   } catch (err) {
     if (throwError) throw err;else util.justWarning(err.message);
   }