+++ es/core/set-state.js
@@ -1,3 +1,2 @@
 import setState from './helper/set-state';
-console.log('what is setState', setState);
 export default setState;